Make HandleInputCache nested static
